+++ src/routes/account/login/github/+server.ts
@@ -1,7 +1,8 @@
 import { GITHUB_CLIENT_SECRET } from "$env/static/private";
 import { PUBLIC_GITHUB_CLIENT_ID } from "$env/static/public";
 import { linkExternalAccount, loginWithExternalIdentity } from "$lib/authentication";
-import { falsePropNames, log } from "$lib/util";
+import { decrypt } from "$lib/crypto";
+import { decodeB64URL, falsePropNames, log } from "$lib/util";
 import { error, redirect } from "@sveltejs/kit";
 import type { LinkAccountButtonOptions, SignInButtonOptions } from "src/types/types";
 import type { RequestHandler } from "./$types";
@@ -10,6 +11,7 @@ export const prerender = false;
 
 async function getProfileFromGithubCallback(url: URL) {
 	const code = <string>url.searchParams.get("code");
+	console.log(`getting access token from code ${code}`);
 
 	const atParams = new URLSearchParams({
 		client_id: PUBLIC_GITHUB_CLIENT_ID,
@@ -25,8 +27,10 @@ async function getProfileFromGithubCallback(url: URL) {
 	}).then((res) => res.json());
 
 	const { access_token } = atResponse;
+	console.log(`got access token ${access_token}, getting profile info`);
 
 	const res = await fetch("https://api.github.com/user", { headers: { Authorization: `token ${access_token}` } }).then((res) => res.json());
+	console.log(`got profile info ${JSON.stringify(res)}`);
 
 	return {
 		email: res.email,
@@ -41,27 +45,27 @@ export const GET: RequestHandler = async ({ url, locals: { userID, sessionID } }
 	type FWDData = SignInButtonOptions | LinkAccountButtonOptions;
 
 	const { email, picture, name, githubID } = await getProfileFromGithubCallback(url);
-	const { referer, opts: forwardedData }: { opts: FWDData; referer: string } = JSON.parse(<string>url.searchParams.get("options"));
-	const fpn = falsePropNames({ githubID, email, name, picture, forwardedData, referer });
+	const options: FWDData = JSON.parse(decodeB64URL(<string>url.searchParams.get("state")!));
+	const fpn = falsePropNames({ githubID, email, name, picture, options });
 	if (fpn.length) throw error(400, `invalid request: ${fpn} are undefined`);
-	else log({ githubID, email, name, picture, forwardedData, referer });
+	else log({ githubID, email, name, picture, options });
 
-	if ("rememberMe" in forwardedData) {
+	if ("rememberMe" in options) {
 		return loginWithExternalIdentity({
 			idFieldName: "githubID",
 			idValue: githubID,
-			rememberMe: forwardedData.rememberMe,
+			rememberMe: options.rememberMe,
 			profileData: {
 				email,
 				name,
 				picture,
 			},
-			redirect: referer,
+			redirect: new URL(options.referer, url).href,
 		});
 	} else {
 		if (userID && sessionID) {
 			await linkExternalAccount({ idFieldName: "githubID", idValue: githubID, userID, sessionID });
-			throw redirect(303, referer);
+			throw redirect(303, options.referer);
 		} else {
 			throw error(500, `userID or sessionID is undefined: ${JSON.stringify({ sessionID: sessionID, userID: userID })}`);
 		}